Reorganize items in databases directory
