+++ mathesar_ui/src/pages/database/schemas/SchemaListSkeleton.svelte
