+++ mathesar_ui/src/pages/database/schemas/SchemaRow.svelte
