+++ mathesar_ui/src/pages/database/settings/SettingsSection.svelte
@@ -1,0 +1,3 @@
+<script lang="ts">
+
+</script>
